fix: resolved some bugs present
