+++ src/routes/user.routes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const verify = require('../authorization/auth');
+const reqAuth = require('../authorization/reqAuth');
 const { validateUserJoi, validateUpdateJoi } = require('../middlewares/joi');
 const router = express.Router();
 const {
@@ -14,10 +14,10 @@ const {
 router.post('/users', validateUserJoi, createAUser);
 router.get('/users', fetchAllUser);
 router.get('/users/:id/posts', fetchAllpostByUserName);
-router.delete('/users/:id', verify, DeleteAUser);
+router.delete('/users/:id', reqAuth, DeleteAUser);
 router.get('/users/:id', fetchAUser);
 router.put('/users/:id', validateUpdateJoi, editAUser);
-router.get('/docs', (req, res) => {
+router.use('/docs', (req, res) => {
     res.redirect('https://www.google.com');
 });
